Require ActiveResource 2.3.5 or better.
Versions of ActiveResource prior to 2.3.5 have a bug that prevents the
AR object from having its errors populated if the content-type of the
response has any extended information in it. The Tracker API returns a
content-type of "application/xml; charset=utf-8". Because of this, the
story object never gets its errors populated.

https://rails.lighthouseapp.com/projects/8994/tickets/3157-activeresource-doesnt-parse-validations-when-extended-content-type